status package renamed to entitystatus, updated tasklist navigator to check completed tag
