+++ app/controllers/ClassBeneficiaryDescriptionController.scala
@@ -23,7 +23,7 @@ import models.Mode
 import models.Status.Completed
 import navigation.Navigator
 import pages.ClassBeneficiaryDescriptionPage
-import pages.status.ClassBeneficiaryStatus
+import pages.entitystatus.ClassBeneficiaryStatus
 import play.api.data.Form
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
