+++ app/controllers/IndividualBeneficiaryAnswersController.scala
@@ -21,7 +21,7 @@ import javax.inject.Inject
 import models.NormalMode
 import models.Status.Completed
 import navigation.Navigator
-import pages.status.IndividualBeneficiaryStatus
+import pages.entitystatus.IndividualBeneficiaryStatus
 import pages.{IndividualBeneficiaryAnswersPage, IndividualBeneficiaryNamePage, TrusteesAnswerPage}
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
