+++ test/controllers/AddABeneficiaryControllerSpec.scala
@@ -22,7 +22,7 @@ import models.Status.Completed
 import models.{AddABeneficiary, FullName, NormalMode}
 import navigation.{FakeNavigator, Navigator}
 import pages.{AddABeneficiaryPage, ClassBeneficiaryDescriptionPage, IndividualBeneficiaryNamePage}
-import pages.status.{ClassBeneficiaryStatus, IndividualBeneficiaryStatus}
+import pages.entitystatus.{ClassBeneficiaryStatus, IndividualBeneficiaryStatus}
 import play.api.inject.bind
 import play.api.mvc.Call
 import play.api.test.FakeRequest
