+++ test/controllers/AddATrusteeControllerSpec.scala
@@ -21,7 +21,7 @@ import forms.AddATrusteeFormProvider
 import models.Status.Completed
 import models.{AddATrustee, FullName, IndividualOrBusiness, NormalMode, UserAnswers}
 import navigation.{FakeNavigator, Navigator}
-import pages.status.TrusteeStatus
+import pages.entitystatus.TrusteeStatus
 import pages.{TrusteeIndividualOrBusinessPage, TrusteesNamePage}
 import play.api.inject.bind
 import play.api.mvc.Call
